# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AI Gateway API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: ai
    description: AI generation endpoints
  - name: rules
    description: Gateway rules and settings
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /ai/models:
    get:
      operationId: listModels
      tags: [ai]
      summary: List all available endpoints and default models
      security:
        - ApiKeyAuth: []
      responses:
        "200":
          description: Model list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModelList"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /ai/text:
    post:
      operationId: generateText
      tags: [ai]
      summary: Generate text from a prompt
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TextRequest"
      responses:
        "200":
          description: Generated text
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TextResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Blocked by rules
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /ai/chat:
    post:
      operationId: chatCompletion
      tags: [ai]
      summary: Multi-turn chat with an AI model
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatRequest"
      responses:
        "200":
          description: Chat response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Blocked by rules
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /ai/code:
    post:
      operationId: generateCode
      tags: [ai]
      summary: Generate or complete code
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CodeRequest"
      responses:
        "200":
          description: Generated code
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CodeResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Blocked by rules
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /ai/image:
    post:
      operationId: generateImage
      tags: [ai]
      summary: Generate an image from a prompt (returns base64)
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImageRequest"
      responses:
        "200":
          description: Generated image as base64
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImageResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Blocked by rules
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /ai/rules:
    get:
      operationId: getRules
      tags: [rules]
      summary: Get current gateway rules and settings
      security:
        - ApiKeyAuth: []
      responses:
        "200":
          description: Current rules
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GatewayRules"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateRules
      tags: [rules]
      summary: Update gateway rules and settings
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRulesBody"
      responses:
        "200":
          description: Updated rules
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GatewayRules"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key

  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    ChatMessage:
      type: object
      properties:
        role:
          type: string
          enum: [system, user, assistant]
        content:
          type: string
      required:
        - role
        - content

    TextRequest:
      type: object
      properties:
        prompt:
          type: string
        model:
          type: ["string", "null"]
        max_tokens:
          type: ["integer", "null"]
      required:
        - prompt

    TextResponse:
      type: object
      properties:
        response:
          type: string
        model:
          type: string
      required:
        - response
        - model

    ChatRequest:
      type: object
      properties:
        messages:
          type: array
          items:
            $ref: "#/components/schemas/ChatMessage"
        model:
          type: ["string", "null"]
        max_tokens:
          type: ["integer", "null"]
      required:
        - messages

    ChatResponse:
      type: object
      properties:
        response:
          type: string
        model:
          type: string
        message:
          $ref: "#/components/schemas/ChatMessage"
      required:
        - response
        - model
        - message

    CodeRequest:
      type: object
      properties:
        prompt:
          type: string
        language:
          type: ["string", "null"]
        model:
          type: ["string", "null"]
        max_tokens:
          type: ["integer", "null"]
      required:
        - prompt

    CodeResponse:
      type: object
      properties:
        response:
          type: string
        model:
          type: string
        language:
          type: string
      required:
        - response
        - model
        - language

    ImageRequest:
      type: object
      properties:
        prompt:
          type: string
        negative_prompt:
          type: ["string", "null"]
        model:
          type: ["string", "null"]
      required:
        - prompt

    ImageResponse:
      type: object
      properties:
        image_base64:
          type: string
        mime_type:
          type: string
        model:
          type: string
        data_url:
          type: string
      required:
        - image_base64
        - mime_type
        - model
        - data_url

    ModelEndpoint:
      type: object
      properties:
        description:
          type: string
        default_model:
          type: string
      required:
        - description
        - default_model

    ModelList:
      type: object
      properties:
        default_models:
          type: object
          additionalProperties:
            type: string
        authentication:
          type: string
        free_token_url:
          type: string
      required:
        - default_models
        - authentication
        - free_token_url

    GatewayRules:
      type: object
      properties:
        systemPrompt:
          type: string
        blockedWords:
          type: array
          items:
            type: string
        defaultMaxTokens:
          type: integer
        modelText:
          type: string
        modelChat:
          type: string
        modelCode:
          type: string
        modelImage:
          type: string
      required:
        - systemPrompt
        - blockedWords
        - defaultMaxTokens
        - modelText
        - modelChat
        - modelCode
        - modelImage

    UpdateRulesBody:
      type: object
      properties:
        systemPrompt:
          type: ["string", "null"]
        blockedWords:
          type: ["array", "null"]
          items:
            type: string
        defaultMaxTokens:
          type: ["integer", "null"]
        modelText:
          type: ["string", "null"]
        modelChat:
          type: ["string", "null"]
        modelCode:
          type: ["string", "null"]
        modelImage:
          type: ["string", "null"]