# lib/api-zod/src/generated/types/chatRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Gateway API
 * OpenAPI spec version: 0.1.0
 */
import type { ChatMessage } from "./chatMessage";

export interface ChatRequest {
  messages: ChatMessage[];
  /** @nullable */
  model?: string | null;
  /** @nullable */
  max_tokens?: number | null;
}