# lib/api-zod/src/generated/types/codeRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Gateway API
 * OpenAPI spec version: 0.1.0
 */

export interface CodeRequest {
  prompt: string;
  /** @nullable */
  language?: string | null;
  /** @nullable */
  model?: string | null;
  /** @nullable */
  max_tokens?: number | null;
}