# lib/api-zod/src/generated/types/codeResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Gateway API
 * OpenAPI spec version: 0.1.0
 */

export interface CodeResponse {
  response: string;
  model: string;
  language: string;
}