# lib/api-zod/src/generated/types/gatewayRules.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Gateway API
 * OpenAPI spec version: 0.1.0
 */

export interface GatewayRules {
  systemPrompt: string;
  blockedWords: string[];
  defaultMaxTokens: number;
  modelText: string;
  modelChat: string;
  modelCode: string;
  modelImage: string;
}