# lib/api-zod/src/generated/types/modelEndpoint.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Gateway API
 * OpenAPI spec version: 0.1.0
 */

export interface ModelEndpoint {
  description: string;
  default_model: string;
}