# lib/api-zod/src/generated/types/modelList.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Gateway API
 * OpenAPI spec version: 0.1.0
 */
import type { ModelListDefaultModels } from "./modelListDefaultModels";

export interface ModelList {
  default_models: ModelListDefaultModels;
  authentication: string;
  free_token_url: string;
}